1.1

+++ main.py
@@ -1,13 +1,15 @@
-from fastapi import FastAPI, File, UploadFile
+from fastapi import FastAPI, File, UploadFile, HTTPException
 from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
 from fastapi.middleware.cors import CORSMiddleware
+from fastapi.exceptions import RequestValidationError
+from starlette.responses import JSONResponse
+
+from mimetypes import guess_type
 
 from etc.delete_files import delete_files_in_folder
 
-
 app = FastAPI()
-
 
 # Разрешаем доступ с клиента (например, для разработки локально)
 app.add_middleware(
@@ -17,7 +19,6 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-
 
 # Путь для статики
 app.mount("/static", StaticFiles(directory="static"), name="static")
@@ -31,13 +32,38 @@ async def get_home():
     return HTMLResponse(content=html_content)
 
 
-# Маршрут для обработки загружаемых файлов
 @app.post("/upload/")
 async def upload_audio(file: UploadFile = File(...)):
     try:
+        # Проверка MIME-типа
+        mime_type, _ = guess_type(file.filename)
+        if not mime_type or not mime_type.startswith("audio/"):
+            raise HTTPException(
+                status_code=400,
+                detail=f"Файл не является аудиофайлом. "
+            )
+
+        # Очищаем папку и сохраняем новый файл
         delete_files_in_folder("uploads")
-        with open(f"uploads/{file.filename}", "wb") as f:
+        file_path = f"uploads/{file.filename}"
+        with open(file_path, "wb") as f:
             f.write(await file.read())
+
         return {"status": "success", "filename": file.filename}
+    except HTTPException as e:
+        raise e
     except Exception as e:
-        return {"status": "error", "message": str(e)}
+        return {
+            "status": "error",
+            "message": "Произошла ошибка при обработке файла.",
+            "details": str(e),
+        }
+
+
+# Обработчик ошибок валидации
+@app.exception_handler(RequestValidationError)
+async def validation_exception_handler(request, exc):
+    return JSONResponse(
+        status_code=422,
+        content={"status": "error", "message": "Некорректный запрос.", "details": str(exc)},
+    )
